refactor: follow repo convention for files
Rename ButtonWithLoading to button-with-loading
